Revert "Use Lambda bodies"
This reverts commit e017ac3a357e0e5a7d1a2b3f7d0d53e8976f5e39.

+++ src/Qowaiv/Month.cs
@@ -316,7 +316,7 @@ public readonly partial struct Month : IXmlSerializable, IFormattable, IEquatabl
         {
             foreach (var month in All)
             {
-                var m = (byte)(int)month;
+                var m = (byte)month;
                 var full = culture.DateTimeFormat.GetAbbreviatedMonthName(m).Unify();
                 var shrt = culture.DateTimeFormat.GetMonthName(m).Unify();
                 this[culture][full] = m;
